fix: Update go_router to v16 and temporarily disable first initialization check
This commit updates the `go_router` package from `^15.2.4` to `^16.0.0` in `pubspec.yaml` to fix GitHub action fail.

In `lib/app.dart`, the logic for determining the initial route based on `isFirstInitialization` has been temporarily bypassed.
The `routerConfig` now always uses `getRouter(true)`, and the loading indicator previously shown when `isFirstInitialization` was null is commented out. This is a temporary measure, with a "Todo" comment indicating the need to fix an infinite loading issue.