fix: Fix infinite Loading case on Messages screen navigation
The issue is caused by increasing `_messagesSubscriptionCount` in `ChatRepositoryImpl` by re-calling `updateFields` thought the `MessagesViewModel` and `MessagesBody` are not disposed.

It happens when navigating to Messages screen then navigating again to it directly through the browser's back and forward navigation buttons without navigating to another screen.
It could be fixed in multiple ways.